+++ fieldmapper/fieldmapper_test.go
